requirements.txt
Flask==3.1.1
pymongo==4.13.2
beautifulsoup4==4.13.4
requests==2.32.4
fake-useragent==2.2.0
lxml==6.0.0
googletrans==4.0.2
Flask-RESTful==0.3.10
python-dotenv==1.1.1
pyjwt==2.10.1
redis==5.2.1
celery==5.5.3
cryptography==42.0.5